Remove old entity data and replace interface to entity
